fix(desktop): address artifact editing review feedback
Split artifact preview loading, editing, and rendering into focused components and hooks.

Use completed upload count in the artifact query key instead of refetching from an effect.

Generated-By: PostHog Code
Task-Id: 2eb29ff5-ae37-45e0-bdb6-77a1468f5b90
