+++ products/desktop/packages/ui/src/features/sessions/components/ArtifactPreviewContent.tsx
@@ -1,0 +1,296 @@
+import { CrosshairSimpleIcon, XIcon } from "@phosphor-icons/react";
+import type { ResourceComment } from "@posthog/api-client/posthog-client";
+import type {
+  CommentAnchor,
+  CommentTarget,
+} from "@posthog/core/comments/anchors";
+import { Button, Spinner } from "@posthog/quill";
+import { isAllowedImageMimeType } from "@posthog/shared";
+import type { UserBasic } from "@posthog/shared/domain-types";
+import type {
+  Dispatch,
+  ReactElement,
+  ReactNode,
+  RefObject,
+  SetStateAction,
+} from "react";
+import { CodeMirrorEditor } from "../../code-editor/components/CodeMirrorEditor";
+import { DocumentPreviewHeader } from "../../code-editor/components/DocumentPreviewHeader";
+import { MarkdownDocumentPreview } from "../../code-editor/components/MarkdownDocumentPreview";
+import { AnnotatedArtifactHtml } from "./AnnotatedArtifactHtml";
+import { AnnotatedArtifactImage } from "./AnnotatedArtifactImage";
+import { ArtifactDocumentCommentAction } from "./ArtifactDocumentCommentAction";
+import { ArtifactTextAnnotations } from "./ArtifactTextAnnotations";
+import type {
+  CommentLocateRequest,
+  HighlightResolution,
+} from "./commentViewTypes";
+import type {
+  ArtifactPreviewResult,
+  EditableArtifactKind,
+  PreviewData,
+} from "./useArtifactPreviewData";
+
+const SVG_MIME_TYPE = "image/svg+xml";
+
+export function ArtifactPreviewError(): ReactElement {
+  return (
+    <div className="flex h-full items-center justify-center text-muted-foreground">
+      This artifact can’t be previewed.
+    </div>
+  );
+}
+
+function GenericArtifactHeader({
+  name,
+  actions,
+}: {
+  name: string;
+  actions?: ReactNode;
+}): ReactElement {
+  return (
+    <header className="flex h-11 shrink-0 items-center justify-between border-border border-b px-3">
+      <span className="truncate font-[var(--code-font-family)] text-[13px] text-muted-foreground">
+        {name}
+      </span>
+      {actions}
+    </header>
+  );
+}
+
+export function ArtifactPreviewContent({
+  name,
+  taskId,
+  commentTarget,
+  commentsEnabled,
+  canEdit,
+  beginEditing,
+  previewData,
+  previewUrl,
+  showRendered,
+  setShowRendered,
+  commentLoadError,
+  markdownRootRef,
+  markdownContainerRef,
+  annotationComments,
+  focusedThreadId,
+  locateRequest,
+  members,
+  activateThread,
+  createAnchoredComment,
+  onResolutionsChange,
+  imageCommenting,
+  setImageCommenting,
+  onImageError,
+  editableKind,
+  artifactResult,
+}: {
+  name: string;
+  taskId: string;
+  commentTarget: CommentTarget;
+  commentsEnabled: boolean;
+  canEdit: boolean;
+  beginEditing: () => void;
+  previewData: PreviewData | undefined;
+  previewUrl: string | null;
+  showRendered: boolean;
+  setShowRendered: Dispatch<SetStateAction<boolean>>;
+  commentLoadError: ReactNode;
+  markdownRootRef: RefObject<HTMLDivElement | null>;
+  markdownContainerRef: RefObject<HTMLDivElement | null>;
+  annotationComments: ResourceComment[];
+  focusedThreadId: string | null;
+  locateRequest: CommentLocateRequest | null;
+  members: UserBasic[];
+  activateThread: (id: string) => void;
+  createAnchoredComment: (
+    anchor: CommentAnchor,
+    content: string,
+    mentions?: number[],
+  ) => Promise<void>;
+  onResolutionsChange: (resolutions: Map<string, HighlightResolution>) => void;
+  imageCommenting: boolean;
+  setImageCommenting: Dispatch<SetStateAction<boolean>>;
+  onImageError: () => void;
+  editableKind: EditableArtifactKind | null;
+  artifactResult: ArtifactPreviewResult | undefined;
+}): ReactElement {
+  if (typeof previewData === "string") {
+    return (
+      <div className="flex h-full flex-col overflow-hidden">
+        <DocumentPreviewHeader
+          label={name}
+          content={previewData}
+          getContent={() => previewData}
+          showRendered={showRendered}
+          onToggleRendered={() => setShowRendered((rendered) => !rendered)}
+          canEdit={canEdit}
+          onEdit={beginEditing}
+          actions={
+            commentsEnabled ? (
+              <ArtifactDocumentCommentAction
+                target={commentTarget}
+                taskId={taskId}
+              />
+            ) : undefined
+          }
+        />
+        {commentLoadError}
+        {showRendered ? (
+          <div
+            ref={markdownContainerRef}
+            className="relative min-h-0 min-w-0 flex-1 overflow-auto"
+          >
+            <div ref={markdownRootRef}>
+              <MarkdownDocumentPreview
+                content={previewData}
+                components={{ img: () => null }}
+              />
+            </div>
+            {commentsEnabled && (
+              <ArtifactTextAnnotations
+                artifactName={name}
+                rootRef={markdownRootRef}
+                containerRef={markdownContainerRef}
+                comments={annotationComments}
+                activeThreadId={focusedThreadId}
+                locateRequest={locateRequest}
+                members={members}
+                onActivateThread={activateThread}
+                onCreate={createAnchoredComment}
+                onResolutionsChange={onResolutionsChange}
+              />
+            )}
+          </div>
+        ) : (
+          <div className="min-h-0 min-w-0 flex-1 overflow-hidden">
+            <CodeMirrorEditor content={previewData} filePath={name} readOnly />
+          </div>
+        )}
+      </div>
+    );
+  }
+
+  if (
+    previewData &&
+    !(previewData instanceof Blob) &&
+    previewData.kind === "html"
+  ) {
+    return (
+      <div className="flex h-full flex-col overflow-hidden">
+        <DocumentPreviewHeader
+          label={name}
+          content={previewData.html}
+          getContent={() => previewData.html}
+          showRendered
+          canEdit={canEdit}
+          onEdit={beginEditing}
+          actions={
+            commentsEnabled ? (
+              <ArtifactDocumentCommentAction
+                target={commentTarget}
+                taskId={taskId}
+              />
+            ) : undefined
+          }
+        />
+        {commentLoadError}
+        <div className="min-h-0 min-w-0 flex-1">
+          <AnnotatedArtifactHtml
+            html={previewData.html}
+            name={name}
+            commentsEnabled={commentsEnabled}
+            comments={annotationComments}
+            activeThreadId={focusedThreadId}
+            locateRequest={locateRequest}
+            members={members}
+            onActivateThread={activateThread}
+            onCreate={createAnchoredComment}
+            onResolutionsChange={onResolutionsChange}
+          />
+        </div>
+      </div>
+    );
+  }
+
+  if (!previewData) return <ArtifactPreviewError />;
+  if (previewData instanceof Blob && !previewUrl) {
+    return (
+      <div className="flex h-full items-center justify-center">
+        <Spinner />
+      </div>
+    );
+  }
+  if (!previewUrl) return <ArtifactPreviewError />;
+
+  if (
+    previewData instanceof Blob &&
+    (isAllowedImageMimeType(previewData.type) ||
+      previewData.type === SVG_MIME_TYPE)
+  ) {
+    const imageActions = commentsEnabled ? (
+      <div className="flex items-center gap-1">
+        <ArtifactDocumentCommentAction target={commentTarget} taskId={taskId} />
+        <Button
+          size="sm"
+          variant={imageCommenting ? "primary" : "outline"}
+          onClick={() => setImageCommenting((commenting) => !commenting)}
+        >
+          {imageCommenting ? <XIcon /> : <CrosshairSimpleIcon />}
+          {imageCommenting ? "Cancel" : "Pin comment…"}
+        </Button>
+      </div>
+    ) : undefined;
+    return (
+      <div className="flex h-full flex-col overflow-hidden">
+        <GenericArtifactHeader name={name} actions={imageActions} />
+        {commentLoadError}
+        <div className="min-h-0 min-w-0 flex-1">
+          <AnnotatedArtifactImage
+            src={previewUrl}
+            name={name}
+            comments={annotationComments}
+            activeThreadId={focusedThreadId}
+            locateRequest={locateRequest}
+            commenting={commentsEnabled && imageCommenting}
+            members={members}
+            onCommentingChange={setImageCommenting}
+            onActivateThread={activateThread}
+            onCreate={createAnchoredComment}
+            onError={onImageError}
+          />
+        </div>
+      </div>
+    );
+  }
+
+  const documentActions = commentsEnabled ? (
+    <ArtifactDocumentCommentAction target={commentTarget} taskId={taskId} />
+  ) : undefined;
+  return (
+    <div className="flex h-full flex-col overflow-hidden">
+      {editableKind === "plain-text" && artifactResult?.source !== undefined ? (
+        <DocumentPreviewHeader
+          label={name}
+          content={artifactResult.source}
+          getContent={() => artifactResult.source ?? ""}
+          showRendered
+          canEdit={canEdit}
+          onEdit={beginEditing}
+          actions={documentActions}
+        />
+      ) : (
+        <GenericArtifactHeader name={name} actions={documentActions} />
+      )}
+      {commentLoadError}
+      <div className="min-h-0 min-w-0 flex-1">
+        <iframe
+          className="h-full w-full border-0 bg-white"
+          sandbox=""
+          src={previewUrl}
+          title={`Preview of ${name}`}
+        />
+      </div>
+    </div>
+  );
+}
